place image modifier now implemented just need to add images
