+++ inc/modifiers.h
@@ -6,7 +6,7 @@
 /*   By: clovell <clovell@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/29 13:55:52 by kmordaun          #+#    #+#             */
-/*   Updated: 2024/05/03 14:10:56 by kmordaun         ###   ########.fr       */
+/*   Updated: 2024/05/07 20:06:06 by kmordaun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -156,4 +156,5 @@ t_err		modifier_setup(t_list *raw_map_file, t_map *map, t_world *world);
 t_vec2		mod_dir_to_vec2(char dir);
 t_mod		*mod_get_mod(t_map *map, t_modtype type, int *index);
 t_entity_2	*mod_search_name(t_world *world, char *str);
+void		display_placed_img(t_game *game);
 #endif
